% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horno.r
\name{dailyMagnitude}
\alias{dailyMagnitude}
\title{Calculate daily magnitude}
\usage{
dailyMagnitude(
  temperature,
  date,
  reference = c(1981, 2010),
  type = "heat wave"
)
}
\arguments{
\item{temperature}{(vector.numeric) Temperature records.}

\item{date}{(vector.Date) Dates for each record.}

\item{reference}{(vector.integer) Range of years to use as reference for
the thresholds calcualtions.}

\item{type}{(character) Calculation for heat or cold waves.}
}
\value{
(vector.numeric) Daily magntiude
}
\description{
Function that calculates the dialy mangitude of heat/cold wave days
}
\examples{
data(temp_ts)
waves <- dialyMagnitude(temp_ts$temperature, temp_ts$date)

}
\author{
Dante Castro Garro
}
