% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horno.r
\name{dailyThreshold}
\alias{dailyThreshold}
\title{Daily thresholds}
\usage{
dailyThreshold(temperature, date, reference = c(1981, 2010), qthresh = 0.9)
}
\arguments{
\item{temperature}{(vector.numeric) Temperature records.}

\item{date}{(vector.Date) Dates for each record.}

\item{reference}{(vector.integer) Range of years to use as reference for
the thresholds calcualtions.}

\item{qthresh}{(numeric) Quantile to be used as threshold. If not defined,
a default value (0.9 or 0.1) will be used based on the `type` parameter.}
}
\value{
(vector.numeric) threshold for each calendar day
}
\description{
Calculate the calendar thresholds based on a quantile for a 31-day-centered
 window in the reference period.
}
\examples{
data(temp_ts)
ext <- dailyThreshold(temp_ts$temperature, temp_ts$date)

}
\author{
Dante Castro Garro
}
