% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horno.r
\name{extremeDays}
\alias{extremeDays}
\title{Extreme hot/cold days}
\usage{
extremeDays(
  temperature,
  date,
  reference = c(1981, 2010),
  type = "heat wave",
  qthresh = NULL
)
}
\arguments{
\item{temperature}{(vector.numeric) Temperature records.}

\item{date}{(vector.Date) Dates for each record.}

\item{reference}{(vector.integer) Range of years to use as reference for
the thresholds calcualtions.}

\item{type}{(character) Calculation for heat or cold waves.}

\item{qthresh}{(numeric) Quantile to be used as threshold. If not defined,
a default value (0.9 or 0.1) will be used based on the `type` parameter.}
}
\value{
(vector.bool) is it a extreme hot or cold day?
}
\description{
Identification of extreme hot/cold days. An extreme day happens when the
 temperature exceeds a daily threshold.
}
\examples{
data(temp_ts)
ext <- extremeDays(temp_ts$temperature, temp_ts$date)

}
\author{
Dante Castro Garro
}
