% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horno.r
\name{findWaves}
\alias{findWaves}
\title{Finding heat waves}
\usage{
findWaves(temperature, date, reference = c(1981, 2010), type = "heat wave")
}
\arguments{
\item{temperature}{(vector.numeric) Temperature records.}

\item{date}{(vector.Date) Dates for each record.}

\item{reference}{(vector.integer) Range of years to use as reference for
the thresholds calcualtions.}

\item{type}{(character) Calculation for heat or cold waves.}
}
\value{
(data.frame) variables related to the heat or cold waves. date (Date)
same as input. temperature (numeric) same as input. wave_day (bool) is it a
wave day?. wave_event (integer) unique id for each wave event. magnitude_day
(numeric) daily magnitude. magnitude_event (numeric) total magntiude for
each wave event.
}
\description{
General function to find heat waves for a temperature time series. The
function can be used to find cold waves as well.
}
\examples{
data(temp_ts)
waves <- findWaves(temp_ts$temperature, temp_ts$date)

}
\author{
Dante Castro Garro
}
