% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horno.r
\name{groupWaves}
\alias{groupWaves}
\title{Numbering the events}
\usage{
groupWaves(wave_days, date)
}
\arguments{
\item{wave_days}{(vector.bool) is it an extreme wave day?}

\item{date}{(vector.Date) Dates for each record.}
}
\value{
(vector.integer) Unique ID for each heat/cold wave. The rest of the
 days are filled with NA's
}
\description{
Giving uniques ID's to each wave event.
}
\examples{
data(temp_ts)
ext <- extremeDays(temp_ts$temperature, temp_ts$date)
wave_day <- waveDays(ext, temp_ts$date)
wave_event_id <- groupWaves(wave_day, temp_ts$date)

}
\author{
Dante Castro Garro
}
