% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horno.r
\name{magnitudeThresholds}
\alias{magnitudeThresholds}
\title{Mangitude threshold}
\usage{
magnitudeThresholds(temperature, date, reference = c(1981, 2010))
}
\arguments{
\item{temperature}{(vector.numeric) Temperature records.}

\item{date}{(vector.Date) Dates for each record.}

\item{reference}{(vector.integer) Range of years to use as reference for
the thresholds calcualtions.}
}
\value{
(dataframe) 25th and 75th percentiles as data.frame
}
\description{
Calculate the required thresholds for the daily mangitude. The thresholds
 are based on the 25th and 75th percentile of annual maximum temperature.
}
\examples{
data(temp_ts)
waves <- dialyMagnitude(temp_ts$temperature, temp_ts$date)

}
\author{
Dante Castro Garro
}
