% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horno.r
\name{waveDays}
\alias{waveDays}
\title{Wave days}
\usage{
waveDays(extreme_day, date)
}
\arguments{
\item{extreme_day}{(vector.bool) is it an extreme (hot/cold) day?}

\item{date}{(vector.Date) Dates for each record.}
}
\value{
(vector.integer) is it a heat/cold wave day?
}
\description{
Identification of days that belong to a heat/cold wave, which is defined by
 3 or more consecutive extreme days.
}
\examples{
data(temp_ts)
ext <- extremeDays(temp_ts$temperature, temp_ts$date)
wave_day <- waveDays(ext, temp_ts$date)

}
\author{
Dante Castro Garro
}
